version before presentation;

+++ HS.Wpf.ARO/ViewModels/OperationRoomViewModel.cs
@@ -15,6 +15,11 @@ using HS.Wpf.ARO.Models;
 using HS.Wpf.ARO.Messages;
 using HS.Data.Entitites.ARO;
 using System.Windows;
+using System.Threading;
+using ToastNotifications;
+using ToastNotifications.Position;
+using ToastNotifications.Lifetime;
+using ToastNotifications.Messages;
 
 namespace HS.Wpf.ARO.ViewModels
 {
@@ -23,6 +28,8 @@ namespace HS.Wpf.ARO.ViewModels
         private readonly AroUnitOfWork _uow;
         private readonly IMapper _mapper;
         private bool _loaded = false;
+
+        private readonly Notifier _notifier;
 
         public virtual OperationRoomCollectionActionsViewModel CollectionActions { get; set; }
         public virtual OperationRoomActionViewModel SelectedAction { get; set; }
@@ -44,32 +51,52 @@ namespace HS.Wpf.ARO.ViewModels
             var now = DateTime.Now.Date;
             FromDate = new DateTime(now.Year, now.Month, 1);
             ToDate = now;
+
+            _notifier = new Notifier(cfg =>
+            {
+                cfg.PositionProvider = new WindowPositionProvider(
+                    parentWindow: Application.Current.MainWindow,
+                    corner: Corner.TopRight,
+                    offsetX: 10,
+                    offsetY: 150);
+
+                cfg.LifetimeSupervisor = new TimeAndCountBasedLifetimeSupervisor(
+                    notificationLifetime: TimeSpan.FromSeconds(5),
+                    maximumNotificationCount: MaximumNotificationCount.FromCount(5));
+
+                cfg.Dispatcher = Application.Current.Dispatcher;
+            });
         }
 
         public void Save()
         {
             try
             {
-                var data = CollectionActions.Actions.Where(s => s.Model.IsDirty);
-                var list = _mapper.Map<IList<OperationRoomAction>>(data);
-                foreach (var item in list)
+                var models = CollectionActions.Actions.Where(s => s.Model.IsDirty).Select(s => s.Model);
+                OperationRoomAction action;
+                foreach (var model in models)
                 {
-                    if (item.Id > 0)
+                    if (model.Id > 0)
                     {
-                        _uow.OperationRoomRepository.Update(item);
+                        action = _uow.OperationRoomRepository.Entities.Single(s => s.Id == model.Id);
+                        _mapper.Map(model, action);
+                        _uow.OperationRoomRepository.Update(action);
                     }
                     else
                     {
-                        _uow.OperationRoomRepository.Add(item);
+                        action = _mapper.Map<OperationRoomAction>(model);
+                        _uow.OperationRoomRepository.Add(action);
                     }
                 }
                 _uow.Save();
 
+                _notifier.ShowSuccess("Data uložena.");
                 LoadData();
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.Message, "Chyba uložení", MessageBoxButton.OK, MessageBoxImage.Error);
+                _notifier.ShowError("Data se nepocedla uložit.");
             }
         }
 
@@ -82,14 +109,28 @@ namespace HS.Wpf.ARO.ViewModels
         {
             try
             {
-                _uow.OperationRoomRepository.Remove(msg.Id);
-                _uow.Save();
+                var result = MessageBox.Show("Opravdu chcete smazat položku?", "Smazání položky", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                if (result == MessageBoxResult.Yes)
+                {
+                    if (msg.Id > 0)
+                    {
+                        _uow.OperationRoomRepository.Remove(msg.Id);
+                        _uow.Save();
 
-                LoadData();
+                        _notifier.ShowSuccess("Data smazána.");
+                        LoadData();
+                    }
+                    else
+                    {
+                        var entity = CollectionActions.Actions.FirstOrDefault(p => p.Model.Id == msg.Id);
+                        CollectionActions.Actions.Remove(entity);
+                    }
+                }
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.Message, "Chyba smazání", MessageBoxButton.OK, MessageBoxImage.Error);
+                _notifier.ShowError("Smazání položky se nepovedlo.");
             }
         }
 
@@ -100,11 +141,30 @@ namespace HS.Wpf.ARO.ViewModels
 
         public void LoadData()
         {
-            var data = _uow.OperationRoomRepository.Entities.ToList();
-            var list = _mapper.Map<IList<OperationRoomActionViewModel>>(data);
+            try
+            {
+                if (CollectionActions != null && CollectionActions.Actions != null && CollectionActions.Actions.Any(s => s.Model.IsDirty))
+                {
+                    var result = MessageBox.Show("Nemáte uložená data. Chcete pokračovat", "Nemáte uložená data", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                    if (result == MessageBoxResult.No) return;
+                }
 
-            CollectionActions.LoadData(list);
-            Stats.Load(list.Select(s => s.Model).ToList());
+                var data = _uow.OperationRoomRepository.Entities
+                            .Where(p => p.IssueDate >= FromDate && p.IssueDate <= ToDate)
+                            .Where(p => !p.IsDeleted)
+                            .OrderByDescending(d => d.IssueDate);
+
+                var list = _mapper.Map<IList<OperationRoomActionViewModel>>(data);
+                CollectionActions.LoadData(list);
+                Stats.Load(list.Select(s => s.Model).ToList());
+
+                _notifier.ShowInformation("Data načtena.");
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Chyba načítání dat", MessageBoxButton.OK, MessageBoxImage.Error);
+                _notifier.ShowError("Data se nepovedla načíst.");
+            }
         }
 
         public void Loaded()
